requirements.txt
Django==1.9.4
decorator==4.0.9
django-annoying==0.9.0
factory-boy==2.6.1
fake-factory==0.5.7
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
python-dateutil==2.5.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
